Refactor Team Management Actions and Components for Enhanced Functionality
- Updated CreateEditTeamAction to ensure team details, including logo, are saved correctly during edits.
- Enhanced ReadTeamAction to include Pokémon data retrieval and logging for better debugging.
- Modified TeamController to redirect to the correct team detail route after editing.
- Added a new relationship in the Team model for fetching associated Pokémon.
- Updated TeamForm component to handle team ID dynamically during edits.
- Adjusted TeamDetail component to display Pokémon data instead of draft picks for improved clarity.

+++ app/Modules/Teams/Actions/CreateEditTeamAction.php
@@ -50,14 +50,15 @@ class CreateEditTeamAction
     {
         $team = Team::where('id', $request->team_id)->first();
         $team->name = $request->name;
+        $team->save();
         if ($request->hasFile('logo')) {
-            $oldlogo = $team->logo;
-            Storage::disk('s3-team-logos')->delete($oldlogo);
+            if ($team->logo !== null) {
+                $oldlogo = $team->logo;
+                Storage::disk('s3-team-logos')->delete($oldlogo);
+            }
             $logo = (new TeamLogoUploadAction)->upload($request);
-        } else {
-            $logo = null;
-        }
-        $team->logo = $logo;
+            $team->logo = $logo;
+            }
         $team->save();
         return $team;
     }
